Added dry_run statement to VLA hanning smoothing.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@38561 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ pipeline/pipeline/hifv/tasks/hanning/hanning.py
@@ -36,16 +36,16 @@ class Hanning(basetask.StandardTaskTemplate):
     
     def prepare(self):
 
-        try:
-            self._do_hanningsmooth()
-            LOG.info("Removing original VIS " + self.inputs.vis)
-            shutil.rmtree(self.inputs.vis)
-            LOG.info("Renaming temphanning.ms to " + self.inputs.vis)
-            os.rename('temphanning.ms', self.inputs.vis)
-        except:
-            LOG.warn('Problem encountered with hanning smoothing.')
+        if not self._executor._dry_run:
+            try:
+                self._do_hanningsmooth()
+                LOG.info("Removing original VIS " + self.inputs.vis)
+                shutil.rmtree(self.inputs.vis)
+                LOG.info("Renaming temphanning.ms to " + self.inputs.vis)
+                os.rename('temphanning.ms', self.inputs.vis)
+            except:
+                LOG.warn('Problem encountered with hanning smoothing.')
 
-        
         return HanningResults()
     
     def analyse(self, results):
